separating utils posix windows tests into modules
