+++ stor/obs.py
@@ -2,7 +2,6 @@ import locale
 import posixpath
 import sys
 
-import dxpy
 import six
 from swiftclient.service import SwiftError
 from swiftclient.service import SwiftUploadObject
@@ -10,7 +9,6 @@ from swiftclient.service import SwiftUploadObject
 from stor.base import Path
 from stor.posix import PosixPath
 from stor import utils
-import stor
 
 
 def _delegate_to_buffer(attr_name, valid_modes=None):
@@ -166,11 +164,6 @@ class OBSPath(Path):
             DNAnexusError: A dxpy client error occured.
             RemoteError: A s3 client error occurred.
         """
-        from stor_dx.dx import DXPath
-        if six.PY3 and encoding and encoding not in ('utf-8', 'utf8') and \
-                isinstance(self, DXPath):  # pragma: no cover
-            raise ValueError('For DNAnexus paths in Python 3, encoding is always assumed to be '
-                             'utf-8. Please switch your encoding or Python version')
         return OBSFile(self, mode=mode, encoding=encoding)
 
     def list(self):
@@ -453,25 +446,6 @@ class OBSFile(object):
                 self.flush()
             self._buffer.close()
         self.closed = True
-        # we want to wait_on_close on DXPath only if no error was thrown while writing the file
-        if not any(sys.exc_info()):  # pragma: no cover
-            self._wait_on_close()
-
-    def _wait_on_close(self):
-        # don't want to import stor_dx here because it is only installed as extra
-        # if isinstance(self._path, stor_dx.dx.DXPath):
-            wait_on_close = stor.settings.get()['dx']['wait_on_close']
-            if wait_on_close:
-                from stor_dx import dx
-                with dx._wrap_dx_calls():  # this will raise an error too
-                    f = dxpy.DXFile(dxid=self._path.canonical_resource,
-                                    project=self._path.canonical_project)
-                    try:
-                        f.wait_on_close(wait_on_close)
-                    # ignore timeout because we want client code to deal with it
-                    except dxpy.DXError as e:  # pragma: no cover
-                        if 'timeout' not in str(e):
-                            raise
 
     def flush(self):
         """Flushes the write buffer to the OBS path (if it exists)"""
