+++ stor_s3/tests/test_cli_s3.py
@@ -1,5 +1,4 @@
 from __future__ import print_function
-
 import contextlib
 import os
 import mock
@@ -10,8 +9,6 @@ import six
 
 from stor import cli
 from stor import exceptions
-from stor import settings
-from stor.posix import PosixPath
 from stor_s3 import test
 from stor_s3.s3 import S3Path
