+++ stor_swift/test.py
